initial working version
Linter is working without any support for command line options.
Since foodcritic does not work with stdin, we have only had
success with having SublimeLinter use the current file on disk.
This is not ideal, but is better than nothing. If we can figure
out how to get tempfile_suffix working, we will.